Demo map finish
Map is done with a few asset of tiles sets, improvise the map a bit so it would be more natural to be seen. (P.S : scaling 1 tile because it was ruining the  tile set).

Asset that is added
- New pavement tile
- New yellow grass
- New Cement pavement
- New Dirt and some adjustment
- New Parking lot tile set

- Adjustment on the tile set to make it easier to use (Prefabs).

The map is consider Done and a few more things left to add.